fix: implement extensions with template refactor
